Refactor specs to extract shared contexts
Shared setup contexts:
  * instance methods
  * class methods via 'def self.'
  * class methods via 'class << self'
