+++ spec/memo_wise_spec.rb
@@ -3,105 +3,6 @@
 require "values"
 
 RSpec.describe MemoWise do
-  ##
-  # Shared contexts for setting up for testing:
-  #   * Instance methods
-  #   * Class methods defined via: 'def self.foo ...'
-  #   * Class methods defined via: 'class << self; def foo ...'
-  ##
-
-  shared_context "with context for instance methods" do
-    let(:instance) { class_with_memo.new }
-
-    let(:class_with_memo) do
-      Class.new do
-        prepend MemoWise
-
-        DefineMethodsForTestingMemoWise.define_methods_for_testing_memo_wise(
-          target: self,
-          via: :instance
-        )
-
-        # Counter for calls to a protected method
-        def protected_memowise_method_counter
-          @protected_memowise_method_counter || 0
-        end
-
-        # A memoized protected method - only makes sense as an instance method
-        def protected_memowise_method
-          @protected_memowise_method_counter =
-            protected_memowise_method_counter + 1
-          "protected_memowise_method"
-        end
-        protected :protected_memowise_method
-        memo_wise :protected_memowise_method
-
-        # Counter for calls to class method '.positional_args', see below.
-        def self.class_positional_args_counter
-          @class_positional_args_counter || 0
-        end
-
-        # See: "with class method with same name as memoized instance method"
-        #
-        # Used by spec below to verify that `memo_wise :with_positional_args`
-        # memoizes only the instance method, and not this class method sharing
-        # the same name.
-        def self.with_positional_args(a, b) # rubocop:disable Naming/MethodParameterName
-          @class_positional_args_counter = class_positional_args_counter + 1
-          "class_with_positional_args: a=#{a}, b=#{b}"
-        end
-      end
-    end
-  end
-
-  shared_context "with context for class methods via 'def self.'" do
-    let(:class_with_memo) do
-      Class.new do
-        prepend MemoWise
-
-        DefineMethodsForTestingMemoWise.define_methods_for_testing_memo_wise(
-          target: self,
-          via: :self_dot
-        )
-
-        # Counter for calls to instance method '#with_keyword_args', see below.
-        def instance_with_keyword_args_counter
-          @instance_with_keyword_args_counter || 0
-        end
-
-        # See: "doesn't memoize instance methods when passed self: keyword"
-        #
-        # Used by spec below to verify that `memo_wise self: :with_keyword_args`
-        # memoizes only the class method, and not this instance method sharing
-        # the same name.
-        def with_keyword_args(a:, b:) # rubocop:disable Naming/MethodParameterName
-          @instance_with_keyword_args_counter =
-            instance_with_keyword_args_counter + 1
-          "instance_with_keyword_args_counter: a=#{a}, b=#{b}"
-        end
-      end
-    end
-  end
-
-  shared_context "with context for class methods via scope 'class << self'" do
-    let(:class_with_memo) do
-      Class.new do
-        class << self
-          prepend MemoWise
-
-          DefineMethodsForTestingMemoWise.define_methods_for_testing_memo_wise(
-            target: self,
-            via: :instance
-          )
-        end
-      end
-    end
-  end
-
-  ##
-  # Specs start here: one 'describe' block for each public API method
-  ##
-
   describe "#memo_wise" do
     shared_examples "#memo_wise shared examples" do
       #
